agents_catalog/32-sila2-lab-automation-agent/streamlit_app/requirements.txt
streamlit>=1.28.0
grpcio>=1.50.0
grpcio-tools>=1.50.0
plotly>=5.15.0
pandas>=2.0.0
